\name{create_basis}
\alias{create_basis}
\title{Create a list of basis functions}
\usage{
create_basis(nBasis, type = "Cos", basis.pars)
}
\arguments{
  \item{nBasis}{number of functions to construct}

  \item{type}{type of functions to construc e.g., "Cos",
  "Legendre" (type = "Legendre" has not been implemented
  yet)}

  \item{basis.pars}{parameters used in construction of the
  basis functions}
}
\value{
list of the basis functions
}
\description{
Create a list of basis functions...
}

