\name{eval_curves}
\alias{eval_curves}
\title{evaluates the specified curves}
\usage{
eval_curves(argvals, coef, basis.fns)
}
\arguments{
  \item{argvals}{vector of argument values}

  \item{coef}{vector of matrix of coefficients of the basis
  functions}

  \item{basis.fns}{list of basis functions}
}
\description{
evaluates the specified curves
}

