\name{sim_coef_unif}
\alias{sim_coef_unif}
\title{simulates the coefficients of the basis functions as independend uniform[0,1] variates#'}
\usage{
sim_coef_unif(nbasis, nfuns)
}
\arguments{
  \item{nbasis}{number of basis functions representing each
  curve}

  \item{nfuns}{number of curves}
}
\value{
coef matrix with dimension \code{nfuns x nbasis} containing
the coefficients
}
\description{
simulates the coefficients of the basis functions as
independend uniform[0,1] variates#'
}

